split terminal.h into prompt header
